src/ScikitLearnSignatureClassifier/requirements.txt
pandas~=2.1.1
uvicorn~=0.23.2
fastapi~=0.103.1
joblib~=1.3.2
numpy~=1.26.0
scikit-learn~=1.3.1
scikit-image~=0.21.0
python-multipart~=0.0.6
Pillow~=10.0.1